+++ source/Public/Server/Get-IMSupportedMediaType.ps1
@@ -1,16 +1,16 @@
-﻿function Get-IMServerFeature
+﻿function Get-IMSupportedMediaType
 {
     <#
     .DESCRIPTION
-        Retreives Immich server feature
+        Retreives Immich supported media type
     .PARAMETER Session
         Optionally define a immich session object to use. This is useful when you are connected to more than one immich instance.
 
         -Session $Session
     .EXAMPLE
-        Get-IMServerFeature
+        Get-IMSupportedMediaType
 
-        Retreives Immich server feature
+        Retreives Immich supported media type
     #>
 
     [CmdletBinding()]
@@ -18,7 +18,7 @@
         [Parameter()][ImmichSession]$Session = $null
     )
 
-    InvokeImmichRestMethod -noauth -Method Get -RelativePath '/server-info/features' -ImmichSession:$Session
+    InvokeImmichRestMethod -noauth -Method Get -RelativePath '/server/media-types' -ImmichSession:$Session
 
 }
 #endregion