+++ source/Public/Tag/Set-IMTag.ps1
@@ -17,6 +17,8 @@
         Defines the assets to tag
     .PARAMETER RemoveAssets
         Defines the assets to untag
+    .PARAMETER Color
+        Defines the tag color, acceppts a HEX string ie, #000000
     .EXAMPLE
         Set-IMTag -AddAssets <assetid>
 
@@ -46,7 +48,11 @@
         [Parameter()]
         [ValidatePattern('^[0-9a-fA-F]{8}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{4}\b-[0-9a-fA-F]{12}$')]
         [string[]]
-        $RemoveAssets
+        $RemoveAssets,
+
+        [Parameter()]
+        [string]
+        $Color
     )
 
     BEGIN
@@ -68,13 +74,16 @@
     PROCESS
     {
         $id | ForEach-Object {
-            if ($PSCmdlet.ShouldProcess($id, 'Add assets'))
+            if ($PSCmdlet.ShouldProcess($id, 'Update tag'))
             {
                 if ($PSBoundParameters.Keys -contains 'AddAssets') {
-                    InvokeImmichRestMethod -Method PUT -RelativePath "/tags/$PSitem/assets" -ImmichSession:$Session -Body:@{assetIds = ($AddAssets -as [string[]])}
+                    InvokeImmichRestMethod -Method PUT -RelativePath "/tags/$PSitem/assets" -ImmichSession:$Session -Body:@{ids = ($AddAssets -as [string[]])}
                 }
                 if ($PSBoundParameters.Keys -contains 'RemoveAssets') {
-                    InvokeImmichRestMethod -Method DELETE -RelativePath "/tags/$PSitem/assets" -ImmichSession:$Session -Body:@{assetIds = ($RemoveAssets -as [string[]])}
+                    InvokeImmichRestMethod -Method DELETE -RelativePath "/tags/$PSitem/assets" -ImmichSession:$Session -Body:@{ids = ($RemoveAssets -as [string[]])}
+                }
+                if ($PSBoundParameters.Keys -contains 'Color') {
+                    InvokeImmichRestMethod -Method PUT -RelativePath "/tags/$PSitem" -ImmichSession:$Session -Body:@{color = $Color}
                 }
             }
         }